+++ R/helpers.R
@@ -109,6 +109,10 @@ isBlob <- function(col) {
   inherits(col, "blob")
 }
 
+isDifftime <- function(col) {
+  inherits(col, "difftime")
+}
+
 selectOrExecute <- function(conn, sql, records, returning) {
   if (is.null(returning)) {
     execute(conn, sql)
@@ -216,6 +220,8 @@ quoteRecords <- function(conn, records) {
         col <- format(col)
       } else if (isTime(col)) {
         col <- format(col)
+      } else if (isDifftime(col)) {
+        col <- as.numeric(col, units="secs")
       }
     } else if (isPostgres(conn)) {
       if (isDatetime(col)) {
@@ -231,6 +237,8 @@ quoteRecords <- function(conn, records) {
           # removes AsIs
           col <- blob::as.blob(lapply(col, function(x) { x }))
         }
+      } else if (isDifftime(col)) {
+        col <- as.numeric(col, units="secs")
       }
     } else if (isSQLite(conn)) {
       # since no standard, store dates and datetimes in the same format as Rails
@@ -241,6 +249,8 @@ quoteRecords <- function(conn, records) {
         col <- format(col)
       } else if (isTime(col)) {
         col <- format(col)
+      } else if (isDifftime(col)) {
+        col <- as.numeric(col, units="secs")
       }
     }
     quoted_records[, i] <- dbQuoteLiteral(conn, col)