+++ tests/testthat/helper-is.R
@@ -1,13 +1,13 @@
+isPostgres <- function(conn) {
+  isRPostgreSQL(conn) || isRPostgres(conn) || isODBCPostgres(conn)
+}
+
 isRPostgreSQL <- function(conn) {
   inherits(conn, "PostgreSQLConnection")
 }
 
 isRPostgres <- function(conn) {
   inherits(conn, "PqConnection")
-}
-
-isPostgres <- function(conn) {
-  isRPostgreSQL(conn) || isRPostgres(conn) || isODBCPostgres(conn)
 }
 
 isODBCPostgres <- function(conn) {
@@ -19,7 +19,7 @@ isRMySQL <- function(conn) {
 }
 
 isMySQL <- function(conn) {
-  isRMySQL(conn) || isRMariaDB(conn)
+  isRMySQL(conn) || isRMariaDB(conn) || isODBCMySQL(conn)
 }
 
 isRMariaDB <- function(conn) {
@@ -28,4 +28,12 @@ isRMariaDB <- function(conn) {
 
 isSQLite <- function(conn) {
   inherits(conn, "SQLiteConnection")
+}
+
+isODBCMySQL <- function(conn) {
+  inherits(conn, "MySQL")
+}
+
+isODBC <- function(conn) {
+  !is.null(attr(conn, "info")$odbc.version)
 }