+++ NEWS.md
@@ -5,7 +5,7 @@
 - Added `variables` to `dbxConnect`
 - Added `connect_timeout`, `sslcert`, `sslkey`, and `sslcrl` for RPostgreSQL
 - Prefer `dbx_logging` over `dbx_verbose`
-- Added support for ODBC Postgres
+- Improved support for ODBC
 
 # dbx 0.2.1